Add connectionFromArraySlice
The existing `connectionFromArray` requires an array representing all
the records in a connection. Actually materializing such an array may be
prohibitively expensive when dealing with large connections (containing
thousands of items, or more). This largely defeats the purpose of
pagination, which is supposed to free us from having to deal with total
data sets when subsets suffice.

At the moment, it is possible to do some ghastly workarounds to trick
`connectionFromArray` into thinking it has a fully materialized array.
For example:

```
const paddedItems = new Array(offset).concat(...items);
if (count > offset + args.first) {
  paddedResults.length++;
}
const connection = connectionFromArray(paddedItems, args);
```

In other words, we create an array with, say, 1000 empty slots, then
insert the 10 items actually retrieved, then increment the `length`
property on the array to trick `connectionFromArray` into thinking that
`hasNextPage` should be true.

This commit seeks to address this common use case by implementing
`connectionFromArraySlice`. This effectively allows you to fetch the
data for a subset of a connection, and pass that in with some metadata
indicating where the subsequence starts and how long the total array
would be if materialized, and get a connection object back.

We also add `connectionFromPromisedArraySlice`, for parity with
`connectionFromPromisedArray`, and we reimplement the existing
`connectionFromArray` to call through to the new function.

All specs pass, except one, which I actually think is a bug in the old
implementation, so I changed it. The test case uses `first: 0` with no
`before` or `after`. Previously, the code returned `hasNextPage: false`,
but according to the [spec](), it seems it should actually return
`true`.

[HasNextPage]() is specified as follows:

1. If first was not set, return false.
2. Let edges be the result of calling ApplyCursorsToEdges(allEdges,
   before, after).
3. If edges contains more than first elements, return true.
4. Return false.

Given that there are no `before` or `after` calls, and
[ApplyCursorsToEdges]() specifies that `allEdges` should be returned in
this case, it seems we should be returning `true` as per step 3 above.

One final thing to note: I've exported `cursorToOffset` as it can be
useful to schema authors when they prepare array slices suitable for
passing in to `connectionFromArraySlice`.

[ApplyCursorsToEdges]: https://facebook.github.io/relay/graphql/connections.htm#ApplyCursorsToEdges()
[HasNextPage]: https://facebook.github.io/relay/graphql/connections.htm#HasNextPage()
[spec]: https://facebook.github.io/relay/graphql/connections.htm