Export getOffset, and give it a more descriptive name
In building a schema using `connectionFromArraySlice`, I've found it
useful to have access to functionality like `getOffset`. This commit
exposes it as an export, after changing the name to be more descriptive,
do further distinguish it from the existing `cursorToOffset`.

Here's an example of the function in use:

    resolve: async (_, args) => {
      const offset = getOffsetWithDefault(args.after, -1) + 1;
      const [articles, totalCount] = await articleLoader.readIndex(
        args.first, offset
      );
      return connectionFromPromisedArraySlice(
        articleLoader.loadMany(articles),
        args,
        {
          sliceStart: offset,
          arrayLength: totalCount,
        },
      );
    },

Without the function, I end up reimplementing equivalent logic, such as:

    const after = args.after && cursorToOffset(args.after);
    const offset = isNaN(after) ? 0 : after + 1;