Expose `offsetToCursor` method
We already export a `cursorForObjectInConnection` method which itself
calls `offsetToCursor`, but because it relies internally on
`Array.prototype.indexOf` to derive the offset, it is not well suited
for use cases where objects may be equivalent but not strictly equal
(`===`).

By exporting `offsetToCursor`, schema authors can derive cursors on
their own without having to go through `cursorForObjectInConnection`.
This mirrors what we do with `cursorToOffset`, which is already
exported.

Fixes: https://github.com/graphql/graphql-relay-js/issues/29