Make compatible with graphql-js v0.5.0
The third argument to the `resolve` functions is now a `context`, and
what was formerly the third argument (the info containing `rootValue`
and so on) is now the fourth argument.

See:

https://github.com/graphql/graphql-js/releases/tag/v0.5.0

Also squashes in additional test fixes (additional `types` declarations)
from @taion:

https://github.com/taion/graphql-relay-js/commit/73c5c3fefb9

These are required as noted in the v0.5.0 release notes.

> Types can now be explicitly provided to new GraphQLSchema({ types:
> [...] }). Types which implement an interface but are otherwise not
> referenced as a field return type are no longer automatically added to
> the Schema!