array connection: add totalCount to Connection
The totalCount attribute on connection types makes it easy to get the
size of the entire collection and makes it explicit. This is especially
useful for large collections where calculating the total size can be
expensive. It's already passed in through the arguments anyway in meta,
but this makes sure it is exposed in the Connection object returned.

An example query:

```
{
  empire {
    ships {
      totalCount
    }
  }
}
```